+++ src/library/DIPS.Mobile.UI/Components/BottomSheets/BottomSheet.cs
@@ -96,21 +96,28 @@ namespace DIPS.Mobile.UI.Components.BottomSheets
                 SearchBar.TextChanged -= OnSearchTextChanged;
             }
         }
-    }
 
-    public enum Positioning
-    {
-        /// <summary>
-        /// The medium position which covers half of the screen.
-        /// </summary>
-        Medium = 0,
-        /// <summary>
-        /// A large position which covers most of the screen.
-        /// </summary>
-        Large = 2,
-        /// <summary>
-        /// The position is determined by the content and will fit the screen.
-        /// </summary>
-        Fit = 4,
+        internal Border CreateBottomBar()
+        {
+            var border = new Border()
+            {
+                Padding = Sizes.GetSize(SizeName.size_2),
+                StrokeThickness = 0,
+                VerticalOptions = LayoutOptions.End,
+                HorizontalOptions = LayoutOptions.Center,
+                BackgroundColor = this.BackgroundColor
+            };
+            var horizontalStackLayout = new HorizontalStackLayout()
+            {
+                HorizontalOptions = LayoutOptions.Center, VerticalOptions = LayoutOptions.End,
+            };
+            foreach (var button in BottombarButtons)
+            {
+                horizontalStackLayout.Add(button);
+            }
+
+            border.Content = horizontalStackLayout;
+            return border;
+        }
     }
 }
