+++ src/library/DIPS.Mobile.UI/Components/BottomSheets/Positioning.cs
@@ -1,0 +1,19 @@
+namespace DIPS.Mobile.UI.Components.BottomSheets;
+
+public enum Positioning
+{
+    /// <summary>
+    /// The medium position which covers half of the screen.
+    /// </summary>
+    Medium = 0,
+
+    /// <summary>
+    /// A large position which covers most of the screen.
+    /// </summary>
+    Large = 2,
+
+    /// <summary>
+    /// The position is determined by the content and will fit the screen.
+    /// </summary>
+    Fit = 4,
+}
